refactor(EatInOrder): EatInOrder Bounded Context 기준으로 패키지를 구분하라
