+++ src/main/java/kitchenpos/eatinorder/domain/JpaOrderTableRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.eatinorder.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
