+++ src/main/java/kitchenpos/eatinorder/domain/OrderType.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.eatinorder.domain;
 
 public enum OrderType {
     DELIVERY, TAKEOUT, EAT_IN
