+++ src/main/java/kitchenpos/eatinorder/ui/OrderRestController.java
@@ -1,7 +1,7 @@
-package kitchenpos.ui;
+package kitchenpos.eatinorder.ui;
 
-import kitchenpos.application.OrderService;
-import kitchenpos.domain.Order;
+import kitchenpos.eatinorder.application.OrderService;
+import kitchenpos.eatinorder.domain.Order;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
